LayoutTests/platform/chromium/inspector/console/console-format-collections-expected.txt

CONSOLE MESSAGE: line 15: [object NodeList]
CONSOLE MESSAGE: line 19: [object HTMLCollection]
CONSOLE MESSAGE: line 23: [object HTMLOptionsCollection]
CONSOLE MESSAGE: line 27: undefined
CONSOLE MESSAGE: line 31: [object HTMLCollection]
CONSOLE MESSAGE: line 35: [object NodeList]
CONSOLE MESSAGE: line 41: 1,2,
CONSOLE MESSAGE: line 44: [object Object]
CONSOLE MESSAGE: line 51: [object Arguments]
Tests that console nicely formats HTML Collections and NodeLists.

[<select id="sel" name="sel">…</select>] console-format-collections.html:15
[<script src="../../http/tests/inspector/inspector-test.js"></script>, <script src="../../http/tests/inspector/console-test.js"></script>, <script>…</script>] console-format-collections.html:19
[<option value="1">one</option>, <option value="2">two</option>] console-format-collections.html:23
[<html>…</html>, <head>…</head>, <script src="../../http/tests/inspector/inspector-test.js"></script>, <script src="../../http/tests/inspector/console-test.js"></script>, <script>…</script>, <body onload="onload()">…</body>, <p>
Tests that console nicely formats HTML Collections and NodeLists.
</p>, <div style="display:none">…</div>, <form id="f">…</form>, <select id="sel" name="sel">…</select>, <option value="1">one</option>, <option value="2">two</option>, <input type="radio" name="x" value="x1">, <input type="radio" name="x" value="x2">] console-format-collections.html:27
[<select id="sel" name="sel">…</select>, <input type="radio" name="x" value="x1">, <input type="radio" name="x" value="x2">] console-format-collections.html:31
[<input type="radio" name="x" value="x1">, <input type="radio" name="x" value="x2">] console-format-collections.html:35
[1, Array[2]] console-format-collections.html:41
NonArrayWithLength console-format-collections.html:44
[1, "2"] console-format-collections.html:51
